Add toggle button component
